+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/UserScoreService.java
@@ -1,6 +1,7 @@
 package com.itachallenge.gamification.service;
 
 import com.itachallenge.challenge.dto.gamification.PointsHistoryResponseDto;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.UUID;
@@ -8,5 +9,7 @@ import java.util.UUID;
 public interface UserScoreService {
 
     Mono<PointsHistoryResponseDto> getUserPointsHistory(UUID userId);
+
+    Flux<Object> recordPoints(UUID userUuid, UUID challengeUuid, int pointsOnSubmissionComplete);
 }